+++ src/Instance/DnDContext.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const DnDContext = React.createContext();
+
+export default DnDContext;