+++ src/Instance/HoldingsList/Holding/HoldingContainer.js
@@ -1,10 +1,84 @@
 import React, {
   useCallback,
+  useMemo,
 } from 'react';
+import ReactDOM from 'react-dom';
 import PropTypes from 'prop-types';
 import { withRouter } from 'react-router';
+import { Draggable } from 'react-beautiful-dnd';
+import { FormattedMessage } from 'react-intl';
 
 import Holding from './Holding';
+
+const dragStyles = {
+  background: '#333',
+  color: '#fff',
+  borderRadius: '3px',
+  padding: '5px',
+  width: 'fit content',
+  height: 'fit content'
+};
+
+const getItemStyle = (draggableStyle, dragging) => {
+  return {
+    userSelect: 'none',
+    width: '100%',
+    ...draggableStyle,
+    ...(dragging ? dragStyles : {}),
+  };
+};
+
+const DraggableHolding = ({
+  provided,
+  snapshot,
+  draggingHoldingsCount,
+  holding,
+  referenceData,
+  onViewHolding,
+  onAddItem,
+  ...rest
+}) => {
+  const rowStyles = useMemo(() => (
+    getItemStyle(
+      provided.draggableProps.style,
+      snapshot.isDragging,
+    )
+  ), [provided.draggableProps.style, snapshot.isDragging]);
+  const child = (
+    <div
+      id={`item-row-${holding.id}`}
+      data-row-inner
+      ref={provided.innerRef}
+      {...provided.draggableProps}
+      {...provided.dragHandleProps}
+
+      style={rowStyles}
+    >
+      {
+        snapshot.isDragging
+          ? (
+            <FormattedMessage
+              id="ui-inventory.moveItems.move.holdings.count"
+              values={{ count: draggingHoldingsCount || 1 }}
+            />
+          ) : (
+            <Holding
+              {...rest}
+              isDraggable
+              holding={holding}
+              referenceData={referenceData}
+              onViewHolding={onViewHolding}
+              onAddItem={onAddItem}
+            />
+          )
+      }
+    </div>
+  );
+
+  if (!snapshot.isDragging) return child;
+
+  return ReactDOM.createPortal(child, document.getElementById('ModuleContainer'));
+};
 
 const HoldingContainer = ({
   location,
@@ -13,6 +87,9 @@ const HoldingContainer = ({
   instance,
   holding,
   referenceData,
+  isDraggable,
+  holdingindex,
+  draggingHoldingsCount,
   ...rest
 }) => {
   const onViewHolding = useCallback(() => {
@@ -29,7 +106,26 @@ const HoldingContainer = ({
     });
   }, [instance.id, holding.id]);
 
-  return (
+  return isDraggable ? (
+    <Draggable
+      key={`${holding.id}`}
+      draggableId={`holding-${holding.id}`}
+      index={holdingindex}
+    >
+      {(provided, snapshot) => (
+        <DraggableHolding
+          provided={provided}
+          snapshot={snapshot}
+          draggingHoldingsCount={draggingHoldingsCount}
+          holding={holding}
+          referenceData={referenceData}
+          onViewHolding={onViewHolding}
+          onAddItem={onAddItem}
+          {...rest}
+        />
+      )}
+    </Draggable>
+  ) : (
     <Holding
       {...rest}
       holding={holding}
@@ -47,6 +143,8 @@ HoldingContainer.propTypes = {
   instance: PropTypes.object.isRequired,
   holding: PropTypes.object.isRequired,
   referenceData: PropTypes.object.isRequired,
+  holdingindex: PropTypes.number.isRequired,
+  isDraggable: PropTypes.bool,
 };
 
 export default withRouter(HoldingContainer);